mvp-05 dazi/socialapp/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
alembic==1.13.2
redis==5.0.8
